Refactoring of packages introduced. Currently hardcoded to put unassigned player into a given game. Progress made
